+++ src/Events/MessageFallbackCreatedEvent.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Topoff\Messenger\Events;
 
 use Illuminate\Foundation\Events\Dispatchable;