+++ src/Models/Message.php
@@ -62,8 +62,10 @@ class Message extends Model
 {
     use DateScopesTrait, SoftDeletes;
 
+    #[\Override]
     public $timestamps = false;
 
+    #[\Override]
     protected $guarded = [];
 
     public function __construct(array $attributes = [])